Clarify the usage of metals.javaHome
On the main metals repo we've gotten it reported a couple times that the
message about setting java home incorrectly sounds like you should set
the `java.home` setting, which isn't used by the Metals extensions, but
rather the `metals.javaHome` is. This just clarifies that and is a
follow-up to the same thing done in the VS Code extension. https://github.com/scalameta/metals-vscode/pull/644